feat: JWT on all routes

+++ src/api/routes/payroll_routes.py
@@ -1,6 +1,7 @@
 from flask import jsonify, Blueprint, request
 from api.models import db, Employee, Company, Role, Salary, Payroll
 from flask_cors import CORS
+from flask_jwt_extended import get_jwt_identity, jwt_required
 
 payroll_bp = Blueprint('payroll', __name__, url_prefix = '/payroll')
 
@@ -10,14 +11,27 @@ CORS(payroll_bp)
 
 #todos_los_roles
 @payroll_bp.route("/", methods=["GET"])
+@jwt_required()
 def get_all_payrolls():
+    employee_id = int(get_jwt_identity())
+    employee = db.session.get(Employee, employee_id)
+    if not employee:
+        return jsonify({"error": "Employee not found"}), 404
+    
+
     payrolls = Payroll.query.all()
     return jsonify([p.serialize() for p in payrolls]), 200
 
 
 #por_id
 @payroll_bp.route("/<int:id>", methods=["GET"])
+@jwt_required()
 def get_payroll(id):
+    employee_id = int(get_jwt_identity())
+    employee = db.session.get(Employee, employee_id)
+    if not employee:
+        return jsonify({"error": "Employee not found"}), 404
+    
     payroll = db.session.get(Payroll, id)
     if not payroll:
         return jsonify({"error": "Payroll not found"}), 400
@@ -25,7 +39,14 @@ def get_payroll(id):
 
 
 @payroll_bp.route("/", methods=["POST"])
+@jwt_required()
 def create_payroll():
+    employee_id = int(get_jwt_identity())
+    employee = db.session.get(Employee, employee_id)
+    if not employee:
+        return jsonify({"error": "Employee not found"}), 404
+    
+
     data = request.get_json(silent=True)
     if not data:
         return jsonify({"error": "JSON body required"}), 400
@@ -43,7 +64,14 @@ def create_payroll():
 
 
 @payroll_bp.route("/<int:id>", methods=["PUT"])
-def update_payroll(id): 
+@jwt_required()
+def update_payroll(id):
+    employee_id = int(get_jwt_identity())
+    employee = db.session.get(Employee, employee_id)
+    if not employee:
+        return jsonify({"error": "Employee not found"}), 404
+
+
     payroll = db.session.get(Payroll, id)
     if not payroll:
         return jsonify({"error": "Payroll not found"}), 404
@@ -62,12 +90,17 @@ def update_payroll(id):
 
 
 @payroll_bp.route("/<int:id>", methods=["DELETE"])
+@jwt_required()
 def delete_payroll(id):
+    employee_id = int(get_jwt_identity())
+    employee = db.session.get(Employee, employee_id)
+    if not employee:
+        return jsonify({"error": "Employee not found"}), 404
+
     payroll = db.session.get(Payroll, id)
     if not payroll:
         return jsonify({"error": "Payroll not found"}), 404
-    
 
     db.session.delete(payroll)
     db.session.commit()
-    return jsonify({"message": "Payroll succesfully deleted"}), 200
+    return jsonify({"message": "Payroll successfully deleted"}), 200
